Dog subclass specifies species for user
Since we know the species is "dog" for the Dog subclass of Pet, instead of having the user specify species='dog' in params, we will just have the Dog class fill that in for the user, simplifying the interface.   I also moved the 'breed' parameter in front of 'name', to fit a logical heirarchy - species -> breed -> name.